+++ locales/en.json
@@ -89,10 +89,8 @@
   "lowestPrice": "Lowest Price",
   "royalty": "royalty",
   "Instance_tooltip": "Instance is like the identifier of an NFT, like a marketplace ticker. It's a \"short computer-friendly name\"",
-  "Minting Settings": "Minting Settings",
+  "mintingSettings": "Minting Settings",
   "user interface mode": "User Interface Mode",
-  "advanced user interface": "Advanced User Interface",
-  "basic user interface": "Basic User Interface",
   "Layout Options": "Layout Options",
   "Gallery Items Per Page": "Gallery Items Per Page",
   "Collections Per Page": "Collections Per Page",
@@ -1143,19 +1141,8 @@
     "bridging": "Bridging {0} {1}"
   },
   "preferences": {
-    "basic": "basic user interface",
-    "advanced": "advanced user interface",
-    "theater": "Enable Theatre View on default",
-    "fold": {
-      "history": "Fold history and price chart in NFT view on default",
-      "description": "Fold description in collection view on default"
-    },
-    "yolo": "Replace BUY NOW with YOLO",
-    "priceVisible": "Always show price in gallery and collection view by default",
-    "mintTimeVisible": "Always show mint time in collection view by default",
-    "loadAllArtworks": "Load all artworks in gallery on default (videos, etc.)",
-    "enableGyroEffect": "Enable gyroscopic effect in gallery for 3d NFTs",
-    "exploreTabOrder": "Default tab in explore"
+    "helpCoverCosts": "Help Cover Costs",
+    "userData": "User Data"
   },
   "navbar": {
     "teleport": "Teleport",
