+++ locales/ru.json
@@ -38,10 +38,8 @@
   "price": "Цена",
   "royalty": "Роялти",
   "Instance_tooltip": "Экземпляр - это короткий идентификатор NFT, тикер на маркете. Это \"короткое компьютерное название\"",
-  "Minting Settings": "Настройки чеканки",
+  "mintingSettings": "Настройки чеканки",
   "user interface mode": "Режим пользовательского интерфейса",
-  "advanced user interface": "Расширенный пользовательский интерфейс",
-  "basic user interface": "Базовый пользовательский интерфейс",
   "Layout Options": "Параметры макета",
   "Gallery Items Per Page": "Элементы галереи на странице",
   "Collections Per Page": "Коллекций на странице",
@@ -727,19 +725,8 @@
     "donationSentTo": "Ваше пожертвование будет отправлено по адресу:"
   },
   "настройки": {
-    "basic": "базовый пользовательский интерфейс",
-    "advanced": "расширенный пользовательский интерфейс",
-    "theater": "Включить просмотр театра по умолчанию",
-    "составлять": {
-      "history": "История составлений и график цен в просмотре NFT по умолчанию",
-      "description": "Свернуть описание в просмотре коллекции по умолчанию"
-    },
-    "yolo": "Замените КУПИТЬ СЕЙЧАС на YOLO",
-    "priceVisible": "Всегда показывать цену в галерее и коллекции по умолчанию",
-    "mintTimeVisible": "Всегда показывать время чеканки в просмотре коллекции по умолчанию",
-    "loadAllArtworks": "Загружать все произведения искусства в галерее по умолчанию (видео и т.п.)",
-    "enableGyroEffect": "Включить гироскопический эффект в галерее для 3d NFT",
-    "exploreTabOrder": "Вкладка по умолчанию в обзоре"
+    "helpCoverCosts": "Помогите покрыть расходы",
+    "userData": "данные пользователя"
   },
   "навигационная панель": {
     "teleport": "Телепорт"
