+++ locales/ua.json
@@ -37,10 +37,8 @@
   "price": "ціна",
   "royalty": "Роялті",
   "Instance_tooltip": "Instance - це ідентифікатор NFT,схожий на ринковий тікер. Це \"коротенька підказка\"",
-  "Minting Settings": "Налаштування карбування",
+  "mintingSettings": "Налаштування карбування",
   "user interface mode": "Режим інтерфейсу користувача",
-  "advanced user interface": "Розширений інтерфейс користувача",
-  "basic user interface": "Основний інтерфейс користувача",
   "Layout Options": "Параметри макета",
   "Gallery Items Per Page": "Елементи галереї на сторінці",
   "Collections Per Page": "Колекції на сторінці",
@@ -718,17 +716,7 @@
   "налаштування": {
     "basic": "базовий інтерфейс користувача",
     "advanced": "розширений інтерфейс користувача",
-    "theater": "Увімкнути перегляд театру за умовчанням",
-    "складати": {
-      "history": "Історія складань і графік цін у перегляді NFT за умовчанням",
-      "description": "Згорнути опис у перегляді колекції за умовчанням"
-    },
-    "yolo": "Замініть КУПИТИ ЗАРАЗ на YOLO",
-    "priceVisible": "Завжди показувати ціну в галереї та колекції за умовчанням",
-    "mintTimeVisible": "Завжди показувати час карбування у перегляді колекції за умовчанням",
-    "loadAllArtworks": "Завантажувати всі твори мистецтва в галереї за умовчанням (відео тощо)",
-    "enableGyroEffect": "Увімкнути гіроскопічний ефект у галереї для 3d NFT",
-    "exploreTabOrder": "Вкладка за умовчанням у огляді"
+    "theater": "Увімкнути перегляд театру за умовчанням"
   },
   "навігаційна панель": {
     "teleport": "Телепорт"
