+++ pages/settings.vue
@@ -1,38 +1,10 @@
 <template>
-  <section>
-    <NeoTabs v-model="activeTab">
-      <NeoTabItem label="Metadata" value="metadata">
-        <SettingsMetadata />
-      </NeoTabItem>
-      <NeoTabItem label="Interface" value="advanced">
-        <SettingsInterface />
-      </NeoTabItem>
-      <NeoTabItem label="Minting" value="minting">
-        <SettingsMinting />
-      </NeoTabItem>
-      <NeoTabItem label="Keyboard shortcuts" value="keyboardShortcuts">
-        <SettingsKeyboardShortcuts />
-      </NeoTabItem>
-      <NeoTabItem label="User Data" value="userData">
-        <SettingsUserData />
-      </NeoTabItem>
-    </NeoTabs>
-  </section>
+  <Settings />
 </template>
 
 <script lang="ts" setup>
-import { NeoTabItem, NeoTabs } from '@kodadot1/brick'
-
-const route = useRoute()
-const router = useRouter()
-const activeTab = computed({
-  get: () => route.query.tab || 'metadata',
-  set: (value) => {
-    route.query.page = ''
-    router.replace({
-      query: { tab: value },
-    })
-  },
+definePageMeta({
+  layout: 'no-footer',
 })
 
 useHead({
